Safety-gym calibration (#19)
* Initial parameters with stable mse

* Safety-Gym init baseline

* Fix dones and add threshold as parameter to cem

* CEM policy threshold update

* Log losses

* Best so far

* fit statistics bug

* freeze current progression

* Rollout done bug fix

* CEM best score so far bug fix

* Just stuff

* Handpicking bug fixes

* Handpicking some more stuff

* Back to negloglike, sample minor bug fix

* Updates from train by steps batch

* Major improvement in stability

+++ simba/infrastructure/replay_buffer.py
@@ -120,7 +120,7 @@ def add_noise(data, noise_to_signal=0.01):
     mean_data = np.mean(data, axis=0)
     mean_data[mean_data == 0] = 1e-5
     std_of_noise = np.abs(mean_data * noise_to_signal)
-    return data + np.random.normal(0.0, std_of_noise, data.shape)
+    return (data + np.random.normal(0.0, std_of_noise, data.shape)).astype(np.float32)
 
 
 def path_length(path):
